Refactor SysInfoViewModel to use float for usage metrics
Refactored SysInfoViewModel to use float instead of double for CPU, disk, memory, and network usage properties and methods. Updated GetDiskUsage and GetMemoryUsage to return float values and ensured explicit casting in UpdateMetrics. In PerfMonViewModel, initialized Categories in the constructor with sorted names and made it init-only, removing LoadCategories. Updated PerfMonPlusPlusPage.xaml to add violet borders to ListView and StackPanel, with minor formatting improvements.

+++ Vanara-Jnana/Models/SysInfoViewModel .cs
@@ -12,13 +12,13 @@ public class SysInfoViewModel : INotifyPropertyChanged
     private readonly PerformanceCounter _netCounter;
 
     /// <summary>Gets the CPU usage as a percentage of total CPU capacity.</summary>
-    public double CpuUsagePercent { get; private set; }
+    public float CpuUsagePercent { get; private set; }
     /// <summary>Gets the disk usage as a percentage of total disk space.</summary>
-    public double DiskUsagePercent { get; private set; }
+    public float DiskUsagePercent { get; private set; }
     /// <summary>Gets the RAM usage as a percentage of total physical memory.</summary>
-    public double MemoryUsagePercent { get; private set; }
+    public float MemoryUsagePercent { get; private set; }
     /// <summary>Gets the network usage in KB/s.</summary>
-    public double NetworkUsage { get; private set; }
+    public float NetworkUsage { get; private set; }
 
     public event PropertyChangedEventHandler? PropertyChanged;
 
@@ -32,7 +32,7 @@ public class SysInfoViewModel : INotifyPropertyChanged
         _timer.Start();
     }
 
-    private static double GetMemoryUsage()
+    private static float GetMemoryUsage()
     {
         var status = new MEMORYSTATUSEX { dwLength = (uint)Marshal.SizeOf(typeof(MEMORYSTATUSEX)) };
         if (GlobalMemoryStatusEx(ref status))
@@ -47,10 +47,10 @@ public class SysInfoViewModel : INotifyPropertyChanged
         return 0;
     }
 
-    private static double GetDiskUsage()
+    private static float GetDiskUsage()
     {
         var drive = DriveInfo.GetDrives().FirstOrDefault(d => d.IsReady);
-        return drive != null ? 100.0 * (1 - (double)drive.AvailableFreeSpace / drive.TotalSize) : 0;
+        return drive != null ? 100.0f * (1 - (float)drive.AvailableFreeSpace / drive.TotalSize) : 0;
     }
 
     public static string GetPrimaryNetworkInterface()
@@ -63,10 +63,10 @@ public class SysInfoViewModel : INotifyPropertyChanged
 
     private void UpdateMetrics()
     {
-        CpuUsagePercent = Math.Round(_cpuCounter.NextValue(), 1);
-        DiskUsagePercent = Math.Round(GetDiskUsage(), 1);
-        MemoryUsagePercent = Math.Round(GetMemoryUsage(), 1);
-        NetworkUsage = Math.Round(_netCounter.NextValue() / 1024, 1);
+        CpuUsagePercent = (float)Math.Round(_cpuCounter.NextValue(), 1);
+        DiskUsagePercent = (float)Math.Round(GetDiskUsage(), 1);
+        MemoryUsagePercent = (float)Math.Round(GetMemoryUsage(), 1);
+        NetworkUsage = (float)Math.Round(_netCounter.NextValue() / 1024, 1);
 
         OnPropertyChanged(nameof(CpuUsagePercent));
         OnPropertyChanged(nameof(MemoryUsagePercent));
